Solved Warnings And Created Locked Meshes For Start Button & Load Button

+++ SP4/Base/Source/Scenes/Game Flow/Game 4 (Stress)/SceneGame4.h
@@ -85,7 +85,7 @@ private:
 	//vector<CEnemy*> theEnemies;
 	void RenderAIs();
 
-	float Timer;
+	double Timer;
 
 	int ScoreToBeat;
 	int Score;
